+++ events/structs_enum.go
@@ -37,9 +37,11 @@ const (
 	StatusChangeTypeWorkflowStep
 	// StatusChangeTypeAlienTask is a StatusChangeType of type AlienTask
 	StatusChangeTypeAlienTask
+	// StatusChangeTypeAttributeValue is a StatusChangeType of type AttributeValue
+	StatusChangeTypeAttributeValue
 )
 
-const _StatusChangeTypeName = "InstanceDeploymentCustomCommandScalingWorkflowWorkflowStepAlienTask"
+const _StatusChangeTypeName = "InstanceDeploymentCustomCommandScalingWorkflowWorkflowStepAlienTaskAttributeValue"
 
 var _StatusChangeTypeMap = map[StatusChangeType]string{
 	0: _StatusChangeTypeName[0:8],
@@ -49,6 +51,7 @@ var _StatusChangeTypeMap = map[StatusChangeType]string{
 	4: _StatusChangeTypeName[38:46],
 	5: _StatusChangeTypeName[46:58],
 	6: _StatusChangeTypeName[58:67],
+	7: _StatusChangeTypeName[67:81],
 }
 
 func (i StatusChangeType) String() string {
@@ -73,6 +76,8 @@ var _StatusChangeTypeValue = map[string]StatusChangeType{
 	strings.ToLower(_StatusChangeTypeName[46:58]): 5,
 	_StatusChangeTypeName[58:67]:                  6,
 	strings.ToLower(_StatusChangeTypeName[58:67]): 6,
+	_StatusChangeTypeName[67:81]:                  7,
+	strings.ToLower(_StatusChangeTypeName[67:81]): 7,
 }
 
 // ParseStatusChangeType attempts to convert a string to a StatusChangeType